% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARforecastDLM-package.R
\docType{package}
\name{SARforecastDLM-package}
\alias{SARforecastDLM}
\alias{SARforecastDLM-package}
\title{SARforecastDLM: SARforecastDLM}
\description{
one-year survival forecast for wild SSSR Chinook using DLM
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Columbia-Basin-Research-CBR/SARforecastDLM.git}
  \item Report bugs at \url{https://github.com/Columbia-Basin-Research-CBR/SARforecastDLM.git/issues}
}

}
\author{
\strong{Maintainer}: Caitlin OBrien \email{csobrien@uw.edu}

}
\keyword{internal}
